requirements.txt
mysql-connector-python
loguru